+++ src/Preprocessor.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "OptProcessor.hpp"
 #include "Directive.hpp"
-#include <cstdint>
+#include <cstddef>
 #include <vector>
 
 struct File;
@@ -10,7 +10,7 @@ struct PreprocessRes {
 
   // In SOF comments, 1 for a preamble after a singleline, 2 for a preamble after a multiline
   uint8_t prefixNewlineCnt{}; 
-  uintmax_t insertionPos{};
+  size_t insertionPos{};
   std::vector<Directive> directives;
 };
 PreprocessRes preprocess(const Opts& opts, File& file);
